requirements.txt
discord.py
langchain-openai
python-dotenv
openai
